Use grpckt and never allow syslog sniffer to die
Sniffer could die because of ie connection issue with server.
